requirements.txt
Pillow==9.5.0
sympy==1.13.1
synthtiger==1.2.1
torch==2.5.0
torchaudio==2.5.0
torchvision==0.20.0
numpy>=1.20.0
